requirements.txt
blinker==1.9.0
click==8.2.1
Flask==3.0.3
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
Werkzeug==3.1.3
requests==2.32.3
python-dotenv==1.0.1
pytest==8.2.2
